backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
langchain-community
langchain-google-genai
requests
huggingface_hub
